requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
streamlit>=1.28.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
openai>=1.3.0
pymongo>=4.5.0
requests>=2.31.0